servers: vfs, mib, mfs and pfs build, and PRIx64 was wrong for LP64
Eight of the twelve boot-image components now build: ds, rs, pm, sched, vfs,
mib, mfs, pfs - along with libmthread, libminixfs, libbdev, libchardriver,
libblockdriver, libfsdriver and libutil, which they need.

The one worth calling out is <machine/int_fmtio.h>. sys/arch/aarch64 had it
as a single line including the arm version, which is right about everything
except the width of a long: there PRIx64 is "llx", because on 32-bit ARM
int64_t is long long. Here <arm/int_types.h> takes __INT64_TYPE__ from the
compiler and GCC says long, so every ll conversion in the file was one
modifier too wide. Any code printing a dev_t, an off_t or an ino_t through
those macros was wrong, and -Wformat catches it one call at a time. The
aarch64 file is now its own, with ll folded to l wherever the type is 64 bits
or intmax_t; the 8-, 16- and 32-bit macros are unchanged and PTR was already
right.

The rest, by kind:

LP64 truncation. mfs limited s_max_size to LONG_MAX, and the field is an
int32_t, so on LP64 the clamp overflowed to -1 - every file would have looked
unable to grow. libfsdriver stored a pointer in a cp_grant_id_t without going
through an integer type wide enough to hold it. vfs computed the offset of
argv[0] in an int and cast it back to a pointer.

Types that only agreed by coincidence. fsdriver_bread and fsdriver_bwrite are
declared int and defined ssize_t, which is the same type only on ILP32; the
definitions now match the header and the dispatch table. vfs/coredump.c
declared its statics with Elf32_ types and defined them with Elf_ ones - again
identical only on 32 bits.

Array bounds in prototypes, which GCC 15 compares against definitions:
unlink_file in mfs and map_driver in vfs lost bounds their callers never
honoured, mount_fs likewise.

Fixed-width utmp fields. init and libutil fill ut_line, ut_name, ut_host and
ut_id with strncpy, which is exactly right - the fields are NUL-padded and may
be completely full - but indistinguishable from an oversight to the compiler.
Written as memset plus memcpy, they say which they are.

Casts between incompatible function types. init's state machine is a set of
functions returning the next function to run, a type C cannot name; the casts
now go through void *, which states the intent without disabling the warning
for the file. libutil's esetfunc cast exit() into a printf-shaped slot, which
would have been undefined if anything ever called through it - replaced with
a one-line wrapper of the right type.

Two truncation checks earned their place rather than being silenced:
libchardriver and libblockdriver build a DS key from a prefix and a label that
are each DS_MAX_KEYLEN, so a full-length label truncates - and a truncated key
publishes a driver under a name that may belong to another one. They now
refuse instead.

Left: init needs libcrypt, which stops on two more of the same kind; tty and
memory are untried (memory wants the ramdisk image, which is stage 5.5); vm
needs arch/aarch64/pagetable.h, which is stage 5.3.

Kernel and hello-world still build and the kernel still boots to the module
panic.

Classification: [возраст] for the utmp fields, the function-type casts, the
prototype bounds and the truncation checks; [порт] for int_fmtio.h, the LP64
truncations and the coincidental type agreements.

+++ minix/fs/mfs/super.c
@@ -320,9 +320,16 @@ int read_super(struct super_block *sp)
   	return(EINVAL);
   }
 
-  /* Limit s_max_size to LONG_MAX */
-  if ((unsigned long)sp->s_max_size > LONG_MAX) 
-	sp->s_max_size = LONG_MAX;
+  /*
+   * Limit s_max_size to what the field can hold.
+   *
+   * This was written as LONG_MAX, which is that same number only while a
+   * long is 32 bits wide. s_max_size is an int32_t - the on-disk superblock
+   * says so - and on LP64 the assignment overflowed to -1, which would have
+   * made every file look as though it could not grow at all.
+   */
+  if ((unsigned long)sp->s_max_size > INT32_MAX)
+	sp->s_max_size = INT32_MAX;
 
   sp->s_isearch = 0;		/* inode searches initially start at 0 */
   sp->s_zsearch = 0;		/* zone searches initially start at 0 */
